# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Verimaya API
  version: 0.1.0
  description: |
    Multi-tenant B2B SaaS API for health tourism operations (Verimaya).

    **Base URL:** `/v1` (e.g. `http://localhost:3000/v1`)

    **Authentication:**
    - Browser/session routes use better-auth cookies (`SessionGuard` + active organization).
    - Auth endpoints live under `/v1/auth/*` (not repeated here in full).
    - Machine/API access via `Authorization: Bearer vk_...` is planned (`ApiKeyGuard` exists but is not global yet).

    **Errors:** All HTTP errors return `{ error: { code, message }, request_id }`.

    **Pagination:** Cursor-based list endpoints accept `?cursor=` and `?limit=` (1–100, default 25).
    Response includes `next_cursor` (nullable) or `items` + `next_cursor` depending on resource.

    **Idempotency:** Mutating domain endpoints accept optional `Idempotency-Key` header.

    This spec is maintained manually (NestJS + Fastify; no runtime Swagger UI yet).

servers:
  - url: http://localhost:3000/v1
    description: Local development

tags:
  - name: Auth
    description: Session and organization (better-auth)
  - name: Patients
  - name: Contacts
  - name: Webhooks
  - name: Webhook Subscriptions
  - name: WhatsApp
  - name: Reports
  - name: API Keys
  - name: Health

components:
  schemas:
    ApiError:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
        request_id: { type: string }

    Patient:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        full_name: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    Contact:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        display_name: { type: string }
        contact_type_name: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

    InboundMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        chat_name: { type: string, nullable: true }
        sender: { type: string, nullable: true }
        body: { type: string, nullable: true }
        has_media: { type: boolean }
        media_path: { type: string, nullable: true }
        status:
          type: string
          enum: [new, parsed, approved, ignored]
        parsed_records: { type: array, items: { type: object }, nullable: true }
        parse_error: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

    ApiKey:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        key_prefix: { type: string }
        scopes: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time, nullable: true }

    WebhookSubscription:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        url: { type: string, format: uri }
        event_types:
          type: array
          items:
            type: string
            enum: [transaction.created, transaction.updated, patient.created, appointment.created]
        active: { type: boolean }
        created_at: { type: string, format: date-time }

    CursorPagePatients:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Patient' } }
        next_cursor: { type: string, nullable: true }

    CursorPageContacts:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Contact' } }
        next_cursor: { type: string, nullable: true }

    InboxList:
      type: object
      properties:
        messages: { type: array, items: { $ref: '#/components/schemas/InboundMessage' } }
        next_cursor: { type: string, nullable: true }

    AiCorrection:
      type: object
      description: |
        Human correction to an AI-parsed WhatsApp transaction draft (Faz 3 learning signal).
        Append-only — never mutates the original inbox item or transaction.
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        inbound_message_id: { type: string, format: uuid, nullable: true }
        original_parsed: { type: array, items: { type: object } }
        corrected: { type: array, items: { type: object } }
        created_by: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

    CursorPageAiCorrections:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AiCorrection' } }
        next_cursor: { type: string, nullable: true }

    WebhookAccepted:
      type: object
      properties:
        accepted: { type: boolean }
        duplicate: { type: boolean }
        integration_event_id: { type: string, format: uuid }
        inbound_message_id: { type: string, format: uuid }
        job_id: { type: string, format: uuid }

    ReportSummary:
      type: object
      description: Period aggregate (income, expense, net — minor units)
      additionalProperties: true

    ReportMonthly:
      type: object
      description: Monthly income/expense/net buckets (`YYYY-MM`, minor units), aligned with finance-base FX resolution
      additionalProperties: true

    ReportByCategoryDetail:
      type: object
      description: Subtitle breakdown (income/expense/net, minor units) within a single category
      additionalProperties: true

  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    SearchQ:
      name: q
      in: query
      schema: { type: string }

  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: better-auth session cookie (browser clients)
    bearerApiKey:
      type: http
      scheme: bearer
      description: API key `vk_...` (route-level guard; not all endpoints)

paths:
  /health:
    get:
      tags: [Health]
      summary: Liveness
      responses:
        '200':
          description: OK

  /health/ready:
    get:
      tags: [Health]
      summary: Readiness (Postgres + Redis)
      responses:
        '200':
          description: Ready
        '503':
          description: Dependency unavailable

  /me:
    get:
      tags: [Auth]
      summary: Current session user and active organization
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Session payload
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }

  /patients:
    get:
      tags: [Patients]
      summary: List patients (cursor + search)
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SearchQ'
      responses:
        '200':
          description: Patient page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPagePatients' }
    post:
      tags: [Patients]
      summary: Create patient
      security:
        - sessionCookie: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name]
              properties:
                full_name: { type: string }
      responses:
        '201':
          description: Created patient

  /patients/{id}:
    get:
      tags: [Patients]
      summary: Get patient by id
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Patient
        '404':
          description: Not found

  /contacts:
    get:
      tags: [Contacts]
      summary: List contacts (cursor + search)
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SearchQ'
      responses:
        '200':
          description: Contact page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPageContacts' }

  /contacts/{id}:
    get:
      tags: [Contacts]
      summary: Get contact by id
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Contact
        '404':
          description: Not found

  /webhooks/{provider}:
    post:
      tags: [Webhooks]
      summary: Generic queue-first webhook ingest (GHL, Meta, Google, …)
      description: |
        Validates stub signature, writes `integration_events`, enqueues BullMQ job. Returns **202 Accepted**.
        Idempotent on `(provider, external_event_id)`.
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
          example: ghl
        - name: X-Webhook-Signature
          in: header
          required: true
          schema: { type: string }
          description: Must match `WEBHOOK_STUB_SECRET` (dev default `dev-webhook-secret`)
        - name: X-Tenant-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
        - name: X-External-Event-Id
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Accepted for async processing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookAccepted' }
        '401':
          description: Invalid or missing signature

  /webhooks/waha:
    post:
      tags: [Webhooks]
      summary: WAHA WhatsApp webhook ingest
      description: |
        Queue-first ingest for WAHA message events. Writes `inbound_messages`
        (unique on tenant + provider + external_id) and enqueues `inbound_message.process`.
        Returns **202 Accepted**.
      parameters:
        - name: X-Webhook-Signature
          in: header
          required: true
          schema: { type: string }
        - name: X-Tenant-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
        - name: X-External-Event-Id
          in: header
          required: false
          schema: { type: string }
          description: WAHA message id override; otherwise extracted from payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WAHA webhook body (event + payload)
              additionalProperties: true
            example:
              event: message
              session: default
              payload:
                id: true_120363143271144447@g.us_3EB0ABCDEF
                from: 120363143271144447@g.us
                chatName: Fixrav Finans
                body: Sandra 2900 GBP ödeme alındı
                hasMedia: false
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookAccepted' }

  /whatsapp/inbox:
    get:
      tags: [WhatsApp]
      summary: List inbound WhatsApp messages (cursor)
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Inbox page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InboxList' }

  /whatsapp/inbox/{id}:
    get:
      tags: [WhatsApp]
      summary: Get single inbound message
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Inbound message
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InboundMessage' }
        '404':
          description: Not found (or other tenant)

  /whatsapp/inbox/process:
    post:
      tags: [WhatsApp]
      summary: Heuristic parse of every `new` inbox message with text
      description: |
        Skips media-only messages (no text body). Sets each processed message's status to
        `parsed` and stashes `parsed_records` (or `parse_error`) in `payload`.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Batch result
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed: { type: integer }
                  parsed: { type: integer }
                  error: { type: integer }

  /whatsapp/inbox/{id}/parse:
    post:
      tags: [WhatsApp]
      summary: Heuristic parse of a single inbox message
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Parsed transaction drafts (also stashed in the message payload)
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items: { type: object }
        '404':
          description: Not found (or other tenant)

  /whatsapp/inbox/{id}/approve:
    post:
      tags: [WhatsApp]
      summary: Mark inbox message as approved
      description: |
        Does **not** create a transaction — the human confirms via `POST /v1/transactions`
        on the AI aktarım screen; this only records the inbox item's disposition.
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Updated status
        '404':
          description: Not found (or other tenant)

  /whatsapp/inbox/{id}/ignore:
    post:
      tags: [WhatsApp]
      summary: Mark inbox message as ignored
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Updated status
        '404':
          description: Not found (or other tenant)

  /whatsapp/corrections:
    post:
      tags: [WhatsApp]
      summary: Record a human correction to an AI-parsed transaction draft batch
      description: |
        Faz 3 learning signal — stores what the heuristic parser produced (`original_parsed`)
        alongside what the user actually saved (`corrected`). Never mutates the source inbox
        item or transaction; append-only audit trail for future prompt/parser tuning.
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [original_parsed, corrected]
              properties:
                inbound_message_id: { type: string, format: uuid, nullable: true }
                original_parsed: { type: array, items: { type: object } }
                corrected: { type: array, items: { type: object } }
      responses:
        '201':
          description: Correction recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiCorrection' }
    get:
      tags: [WhatsApp]
      summary: List recorded AI corrections (cursor, newest first)
      security:
        - sessionCookie: []
        - bearerApiKey: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Corrections page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPageAiCorrections' }

  /whatsapp/parse:
    post:
      tags: [WhatsApp]
      summary: Heuristic parse stub (no LLM yet)
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
      responses:
        '200':
          description: Parsed transaction drafts

  /reports/summary:
    get:
      tags: [Reports]
      summary: Period income/expense/net summary
      security:
        - sessionCookie: []
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Summary aggregates
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportSummary' }

  /reports/by-category:
    get:
      tags: [Reports]
      summary: Expense/income breakdown by category
      security:
        - sessionCookie: []
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Category breakdown

  /reports/by-category-detail:
    get:
      tags: [Reports]
      summary: Subtitle breakdown within a category
      security:
        - sessionCookie: []
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: category
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Subtitle breakdown
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportByCategoryDetail' }

  /reports/monthly:
    get:
      tags: [Reports]
      summary: Monthly income/expense/net buckets
      security:
        - sessionCookie: []
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Monthly aggregates
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportMonthly' }

  /api-keys:
    get:
      tags: [API Keys]
      summary: List API keys (no hash)
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Key list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ApiKey' }
    post:
      tags: [API Keys]
      summary: Create API key (plaintext returned once)
      security:
        - sessionCookie: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Created (includes `key` field once)

  /api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Revoke API key
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Revoked

  /webhook-subscriptions:
    get:
      tags: [Webhook Subscriptions]
      summary: List outbound webhook subscriptions
      description: |
        Session-only settings CRUD — external API keys never manage a tenant's own
        outbound webhook destinations.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Subscription list (secret never returned)
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/WebhookSubscription' }
    post:
      tags: [Webhook Subscriptions]
      summary: Create outbound webhook subscription
      security:
        - sessionCookie: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, secret, event_types]
              properties:
                url: { type: string, format: uri }
                secret:
                  type: string
                  description: Shared secret used to sign `X-Verimaya-Signature` (never stored in plaintext, never returned)
                  minLength: 16
                event_types:
                  type: array
                  items:
                    type: string
                    enum: [transaction.created, transaction.updated, patient.created, appointment.created]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookSubscription' }

  /webhook-subscriptions/{id}:
    delete:
      tags: [Webhook Subscriptions]
      summary: Delete outbound webhook subscription
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found